make moral_outrage_score int instead of float

+++ schemas/responses.py
@@ -3,5 +3,5 @@ from pydantic import BaseModel
 class MoralOutrage(BaseModel):
     text_id: str
     text: str
-    moral_outrage_score: float
+    moral_outrage_score: int
     label_timestamp: str
